+++ app/src/main/java/com/mskd/flux/data/repository/ddb/DatabaseRepositoryImpl.kt
@@ -3,9 +3,9 @@ package com.mskd.flux.data.repository.ddb
 import com.mskd.flux.data.ddb.DatabaseDao
 import com.mskd.flux.model.UserFile
 import com.mskd.flux.model.artwork.Artwork
-import com.mskd.flux.model.artwork.ContentType
 import com.mskd.flux.model.artwork.Episode
 import com.mskd.flux.model.artwork.Movie
+import com.mskd.flux.model.artwork.Season
 import com.mskd.flux.utils.extensions.tmdbImage
 import com.mskd.flux.utils.extensions.tmdbImageLarge
 import kotlinx.coroutines.flow.Flow
@@ -29,12 +29,20 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
         return dao.flowEpisodes(artworkId = artworkId)
     }
 
+    override fun flowSeasons(artworkId: Long): Flow<List<Season>> {
+        return dao.flowSeasons(artworkId = artworkId)
+    }
+
     override suspend fun saveArtworks(artworks: List<Artwork>) {
         dao.insertArtworks(artworks = artworks)
     }
 
     override suspend fun saveMovies(movies: List<Movie>) {
         dao.insertMovies(movies = movies)
+    }
+
+    override suspend fun saveSeasons(seasons: List<Season>) {
+        dao.insertSeasons(seasons)
     }
 
     override suspend fun saveEpisodes(episodes: List<Episode>) {
@@ -61,10 +69,6 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
         return dao.getMoviesNotInFiles(fileNames =  files.map { it.name })
     }
 
-    override suspend fun getEpisode(episodeId: Long): Episode? {
-        return dao.getEpisode(episodeId = episodeId)
-    }
-
     override suspend fun getEpisodes(artworkId: Long): List<Episode> {
         return dao.getEpisodes(artworkId = artworkId)
     }
@@ -81,6 +85,18 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
         return dao.getEpisodeCountByArtworkId(artworkId = artworkId)
     }
 
+    override suspend fun getEpisodeCountBySeason(artworkId: Long, season: Int): Int {
+        return dao.getEpisodeCountBySeason(artworkId = artworkId, season = season)
+    }
+
+    override suspend fun getSeasons(artworkId: Long): List<Season> {
+        return dao.getSeasons(artworkId).sortedBy { it.season }
+    }
+
+    override suspend fun getSeasons(): List<Season> {
+        return dao.getSeasons()
+    }
+
     override suspend fun getUnknownMedias(): List<Episode> {
         return dao.getUnknownMedias()
     }
@@ -88,25 +104,24 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
     override suspend fun getAllImagesPaths(): List<String> {
         val artworks = dao.getArtworksImages()
         val episodes = dao.getEpisodesImages()
+        val seasons = dao.getSeasonsImages()
 
         return buildList {
             addAll(artworks.filter { it.imagePath.isNotBlank() }.map { it.imagePath.tmdbImage })
             addAll(artworks.filter { it.imagePath.isNotBlank() }.map { it.bannerPath.tmdbImageLarge })
             addAll(episodes.filter { it.isNotBlank() }.map { it.tmdbImage })
+            addAll(seasons.filter { it.isNotBlank() }.map { it.tmdbImage })
         }
     }
 
     override suspend fun deleteArtworks(artworks: List<Artwork>) {
-        dao.deleteArtworks(ids = artworks.map { it.id })
 
-        artworks.distinctBy { it.id }.forEach { artwork ->
+        val artworkIds = artworks.map { it.id }.distinct()
+        dao.deleteArtworks(ids = artworkIds)
+        dao.deleteMoviesByIds(ids = artworkIds)
+        artworkIds.forEach { dao.deleteEpisodesByArtworkId(artworkId = it) }
+        dao.deleteSeasonsByIds(artworkIds = artworkIds)
 
-            when (artwork.type) {
-                ContentType.MOVIE -> dao.deleteMoviesByIds(listOf(artwork.id))
-                ContentType.SHOW -> dao.deleteEpisodesByArtworkId(artworkId = artwork.id)
-            }
-
-        }
     }
 
     override suspend fun deleteMovies(movies: List<Movie>) {
@@ -123,21 +138,9 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
         // Delete episodes
         dao.deleteEpisodesByIds(episodes.map { it.id })
 
-        // Delete related artworks if needed
-        episodes
-            .map { it.artworkId }
-            .distinct()
-            .forEach { artworkId ->
-
-                // Check if it remains episode for show
-                val remainingEpisodes = getEpisodeCount(artworkId = artworkId)
-
-                // If no, delete the show
-                if (remainingEpisodes == 0) {
-                    dao.deleteArtworks(ids = listOf(artworkId))
-                }
-
-            }
+        // Delete empty seasons and artworks
+        dao.deleteEmptySeasons()
+        dao.deleteEmptyArtworks()
 
     }
 
@@ -155,5 +158,6 @@ class DatabaseRepositoryImpl @Inject constructor(private val dao: DatabaseDao) :
         dao.deleteAllArtworks()
         dao.deleteAllMovies()
         dao.deleteAllEpisodes()
+        dao.deleteAllSeasons()
     }
 }