+++ app/src/main/java/com/mskd/flux/data/repository/settings/SettingsRepository.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
 import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.preferencesDataStore
-import com.mskd.flux.ui.theme.Ui
 import kotlinx.coroutines.flow.Flow
 import java.util.Locale
 
@@ -23,8 +22,6 @@ interface SettingsRepository {
 
     suspend fun setPlayerForwardValue(value: Int)
 
-    suspend fun setUiTheme(theme: Ui.THEME)
-
     suspend fun setDataLanguage(locale: Locale?)
 
     suspend fun setSubtitlesLanguage(locale: Locale)
@@ -42,7 +39,6 @@ interface SettingsRepository {
     data class State(
         val playerRewindValue: Int = 10,
         val playerForwardValue: Int = 10,
-        val uiTheme: Ui.THEME = Ui.THEME.SYSTEM,
         val subtitlesLanguage: Locale = Locale.getDefault(),
         val audioLanguage: Locale = Locale.getDefault(),
         val externalPlayer: Boolean = false,