+++ app/src/main/java/com/mskd/flux/model/tmdb/TMDBTranslations.kt
@@ -33,6 +33,13 @@ data class TMDBTranslations(
         val overview: String?
     )
 
+    sealed class Request(val language: Locale) {
+        class Movie(val artworkId: Long, language: Locale) : Request(language)
+        class Show(val artworkId: Long, language: Locale) : Request(language)
+        class Episode(val artworkId: Long, val season: Int, val number: Int, language: Locale) : Request(language)
+        class Season(val artworkId: Long, val season: Int, language: Locale) : Request(language)
+    }
+
 }
 
 fun Collection<TMDBTranslations.Translation>.findWithLocale(locale: Locale) : TMDBTranslations.Translation? {