+++ app/src/main/java/com/mskd/flux/screens/artwork/composables/ArtworkContentLarge.kt
@@ -11,8 +11,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.material3.HorizontalDivider
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -20,12 +19,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.mskd.flux.R
 import com.mskd.flux.mockups.MediaMockups
-import com.mskd.flux.model.artwork.Artwork
-import com.mskd.flux.model.artwork.Episode
+import com.mskd.flux.model.artwork.FullArtwork
 import com.mskd.flux.model.artwork.Media
 import com.mskd.flux.screens.artwork.ArtworkIntent
 import com.mskd.flux.screens.artwork.composables.common.ArtworkButtons
-import com.mskd.flux.screens.artwork.composables.common.ArtworkDescription
+import com.mskd.flux.screens.artwork.composables.common.ArtworkDescriptionsPager
 import com.mskd.flux.screens.artwork.composables.common.ArtworkImage
 import com.mskd.flux.screens.artwork.composables.episodes.EpisodeItem
 import com.mskd.flux.screens.artwork.composables.episodes.SeasonsTabs
@@ -36,9 +34,8 @@ import com.mskd.flux.utils.LandscapePreview
 
 @Composable
 fun ArtworkContentLarge(
-    artwork: Artwork,
-    media: Media,
-    episodes: List<Episode>,
+    fullArtwork: FullArtwork,
+    currentMedia: Media,
     currentSeason: Int,
     scaffoldInnerPadding: PaddingValues,
     sendIntent: (ArtworkIntent) -> Unit,
@@ -52,7 +49,7 @@ fun ArtworkContentLarge(
 
             ArtworkImage(
                 modifier = Modifier.fillMaxSize(),
-                artwork = artwork,
+                fullArtwork = fullArtwork,
                 sendIntent = sendIntent
             )
 
@@ -70,62 +67,85 @@ fun ArtworkContentLarge(
             item {
 
                 ArtworkButtons(
-                    media = media,
+                    media = currentMedia,
                     sendIntent = sendIntent
                 )
 
             }
 
             item {
+                Spacer(modifier = Modifier.height(Ui.Space.LARGE))
+            }
 
-                ArtworkDescription(media = media)
+            item {
+
+                ArtworkDescriptionsPager(
+                    fullArtwork = fullArtwork,
+                    currentMedia = currentMedia
+                )
 
             }
 
-            if (episodes.isNotEmpty()) {
+            item {
+                Spacer(modifier = Modifier.height(Ui.Space.LARGE))
+            }
 
-                item {
+            (fullArtwork as? FullArtwork.FullShow)?.let { show ->
 
-                    Column(
-                        modifier = Modifier.padding(top = Ui.Space.LARGE),
-                        verticalArrangement = Arrangement.spacedBy(Ui.Space.SMALL)
-                    ) {
+                if (show.episodes.isNotEmpty()) {
 
-                        Text.Title.Large(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(horizontal = Ui.Space.MEDIUM),
-                            text = stringResource(R.string.episode_list),
-                            color = MaterialTheme.colorScheme.onBackground
-                        )
+                    item {
+
+                        Column(verticalArrangement = Arrangement.spacedBy(Ui.Space.SMALL)) {
+
+                            Text.Title.Large(
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .padding(horizontal = Ui.Space.MEDIUM),
+                                text = stringResource(R.string.episodes),
+                                emphasized = true,
+                                color = MaterialTheme.colorScheme.onPrimaryContainer
+                            )
+
+                        }
+
+                    }
+
+                    item {
+                        Spacer(modifier = Modifier.height(Ui.Space.MEDIUM))
+                    }
+
+                    item {
 
                         SeasonsTabs(
-                            selectedSeason = currentSeason,
-                            seasons = episodes.map { it.season }.distinct(),
-                            onSeasonTap = { sendIntent(ArtworkIntent.SelectSeason(it)) }
+                            seasons = show.seasons,
+                            currentSeason = currentSeason,
+                            sendIntent = sendIntent
                         )
 
                     }
 
-                }
-
-
-                itemsIndexed(
-                    items = episodes
-                        .filter { it.season == currentSeason }
-                        .sortedBy { it.number },
-                    key = { _, e -> e.id }
-                ) { i, episode ->
-
-                    if (i != 0) {
-                        HorizontalDivider(modifier = Modifier.padding(horizontal = Ui.Space.MEDIUM))
+                    item {
+                        Spacer(modifier = Modifier.height(Ui.Space.MEDIUM))
                     }
 
-                    EpisodeItem(
-                        modifier = Modifier.animateItem(),
-                        episode = episode,
-                        sendIntent = sendIntent
-                    )
+                    items(
+                        items = show.episodes
+                            .filter { it.season == currentSeason }
+                            .sortedBy { it.number },
+                        key = { e -> e.id }
+                    ) { episode ->
+
+                        EpisodeItem(
+                            modifier = Modifier.animateItem(),
+                            episode = episode,
+                            isSelected = episode.id == currentMedia.mediaId,
+                            sendIntent = sendIntent
+                        )
+
+                        Spacer(modifier = Modifier.height(Ui.Space.SMALL))
+
+                    }
 
                 }
 
@@ -147,9 +167,8 @@ fun ArtworkContentLarge(
 fun ArtworkContentLargeMovie_Preview() {
     AppTheme {
         ArtworkContentLarge(
-            artwork = MediaMockups.movieArtwork,
-            media = MediaMockups.movie,
-            episodes = emptyList(),
+            fullArtwork = MediaMockups.fullMovie,
+            currentMedia = MediaMockups.movie,
             currentSeason = -1,
             scaffoldInnerPadding = PaddingValues.Zero,
             sendIntent = {}
@@ -162,9 +181,8 @@ fun ArtworkContentLargeMovie_Preview() {
 fun ArtworkContentLargeShow_Preview() {
     AppTheme {
         ArtworkContentLarge(
-            artwork = MediaMockups.showArtwork,
-            media = MediaMockups.episode1,
-            episodes = MediaMockups.episodesWithStatus,
+            fullArtwork = MediaMockups.fullShow,
+            currentMedia = MediaMockups.episode1,
             currentSeason = 1,
             scaffoldInnerPadding = PaddingValues.Zero,
             sendIntent = {}