+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## Version 1.3.7
+## Version 1.4.0
 
-- Fix: Crash on unknown artworks
+- Feature: Add Seasons
+- Feature: Add artwork description
+- Feature: More colors customization
+- Feature: Full sync on new updates if needed
+- UI: Brand-new artwork screen
+- UI: Progress indicator in full sync
+- Other: New screenshots